split addon code between gui and cli
