+++ src/main/java/de/idrinth/waraddonclient/model/AddonList.java
@@ -1,40 +1,23 @@
 package de.idrinth.waraddonclient.model;
 
 import de.idrinth.waraddonclient.Config;
-import de.idrinth.waraddonclient.Utils;
 import de.idrinth.waraddonclient.service.FileLogger;
 import de.idrinth.waraddonclient.service.Request;
 import de.idrinth.waraddonclient.service.XmlParser;
-import java.awt.event.ActionListener;
-import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashMap;
 import javax.json.JsonArray;
-import javax.swing.JMenu;
-import javax.swing.table.DefaultTableModel;
 
-public class AddonList implements java.lang.Runnable {
+public abstract class AddonList implements Runnable {
 
-    private final HashMap<String, ActualAddon> list = new HashMap<>();
+    protected final HashMap<String, ActualAddon> list = new HashMap<>();
 
-    private final ArrayList<ActualAddon> rows = new ArrayList<>();
-
-    private final HashMap<String, WatchedFile> watchedFilesMap = new HashMap<>();
+    protected final ArrayList<ActualAddon> rows = new ArrayList<>();
     
-    private final Request client;
+    protected final Request client;
     
-    private DefaultTableModel model;
-    
-    private final FileLogger logger;
-
-    private JMenu menu;
-
-    private final HashMap<String, Tag> tags = new HashMap<>();
-
-    private final ArrayList<String> tagNames = new ArrayList<>();
-    
-    private ActionListener listener;
+    protected final FileLogger logger;
     
     private final XmlParser parser;
     
@@ -45,15 +28,6 @@ public class AddonList implements java.lang.Runnable {
         this.logger = logger;
         this.parser = parser;
         this.config = config;
-    }
-
-    public void setMenu(JMenu menu, ActionListener listener) {
-        this.menu = menu;
-        this.listener = listener;
-    }
-
-    public void setModel(DefaultTableModel model) {
-        this.model = model;
     }
 
     public ActualAddon get(int position) {
@@ -83,95 +57,12 @@ public class AddonList implements java.lang.Runnable {
         rows.add(addon);
     }
 
-    /**
-     * returns a list of files watched
-     */
-    public HashMap<String, WatchedFile> getWatchedFiles() {
-        return watchedFilesMap;
-    }
-
-    /**
-     * processes the addon lst to see what addon is tagged with a specific tag
-     */
-    private void processAddons() {
-        for (int counter = 0; counter < size(); counter++) {
-            for (String tag : get(counter).getTags()) {
-                if (!tags.containsKey(tag)) {
-                    tags.put(tag, new Tag(tag, listener));
-                }
-                tags.get(tag).addMember(get(counter));
-                if (!tagNames.contains(tag)) {
-                    tagNames.add(tag);
-                }
-            }
-        }
-    }
-
-    /**
-     * get the selected tags
-     */
-    public java.util.ArrayList<String> getActiveTags() {
-        java.util.ArrayList<String> active = new ArrayList<>();
-        tagNames.stream().filter(tag -> (tags.get(tag).isActive())).forEach(tag -> active.add(tag));
-        return active;
-    }
-
-    /**
-     * removes unneeded tags and adds new tags to the menu
-     */
-    private void processTags() {
-        tagNames.stream().map(name -> {
-            tags.get(name).checkMembers();
-            return name;
-        }).forEach(name -> {
-            if (!tags.get(name).hasMembers()) {
-                menu.remove(tags.get(name).getMenu());
-                tags.remove(name);
-                tagNames.remove(name);
-            } else if (menu != null && tags.get(name).getMenu().getParent() == null) {
-                menu.add(tags.get(name).getMenu());
-            }
-        });
-    }
-
-    /**
-     * method to silently update all addons
-     */
-    public void updateAll() {
-        model = new DefaultTableModel();
-        run();
-        for (int i = 0; i < size(); i++) {
-            Addon addon = get(i);
-            if (addon.getStatus().equals("X")) {
-                try {
-                    addon.install();
-                } catch (Exception ex) {
-                    logger.error(ex);
-                }
-            }
-        }
-    }
-
-    @Override
-    public void run() {
-        try {
-            new JsonProcessor(client.getAddonList(), model).run();
-            processAddons();
-            processTags();
-        } catch (IOException exception) {
-            logger.error(exception);
-        }
-    }
-
     public class JsonProcessor {
 
         private final javax.json.JsonArray json;
-        
-        private final DefaultTableModel model;
 
-        public JsonProcessor(JsonArray parse, DefaultTableModel model) {
+        public JsonProcessor(JsonArray parse) {
             json = parse;
-            this.model = model;
         }
 
         public void run() throws IOException {
@@ -187,21 +78,8 @@ public class AddonList implements java.lang.Runnable {
             }
         }
 
-        /**
-         *
-         * @param model
-         * @param addon
-         * @return javax.swing.table.DefaultTableMode
-         */
-        private void newAddon(ActualAddon addon) {
+        protected void newAddon(ActualAddon addon) {
             add(addon);
-            model.addRow(addon.getTableRow());
-            if (!addon.getFile().isEmpty()) {
-                if (!watchedFilesMap.containsKey(addon.getFile().toLowerCase())) {
-                    watchedFilesMap.put(addon.getFile().toLowerCase(), new WatchedFile());
-                }
-                watchedFilesMap.get(addon.getFile().toLowerCase()).addAddon(addon);
-            }
         }
 
         private void processJsonAddon(de.idrinth.waraddonclient.model.ActualAddon addon) {
@@ -212,56 +90,8 @@ public class AddonList implements java.lang.Runnable {
             newAddon(addon);
         }
 
-        private void existingAddon(ActualAddon addon) {
-            if (!list.containsKey(addon.getName())) {
-                return;
-            }
+        protected void existingAddon(ActualAddon addon) {
             list.get(addon.getName()).update(addon);
-            String[] data = list.get(addon.getName()).getTableRow();
-            model.setValueAt(data[0], rows.indexOf(list.get(addon.getName())), 0);
-            model.setValueAt(data[1], rows.indexOf(list.get(addon.getName())), 1);
-            model.setValueAt(data[2], rows.indexOf(list.get(addon.getName())), 2);
-            model.setValueAt(data[3], rows.indexOf(list.get(addon.getName())), 3);
-        }
-    }
-
-    public class WatchedFile implements java.lang.Runnable {
-
-        private boolean active;
-
-        private File file;
-
-        private final ArrayList<ActualAddon> list = new ArrayList<>();
-
-        /**
-         * Adds an addon to watch an be handled here
-         *
-         * @param addon
-         */
-        public void addAddon(de.idrinth.waraddonclient.model.ActualAddon addon) {
-            list.add(addon);
-        }
-
-        public void setFileToProcess(java.io.File file2process) {
-            while (active) {
-                Utils.sleep(100, logger);
-            }
-            active = true;
-            file = file2process;
-            new java.lang.Thread(this).start();
-        }
-
-        /**
-         * check if a file was actually updated and is avaible
-         */
-        @Override
-        public void run() {
-            if (file == null || !file.exists()) {
-                active = false;
-                return;
-            }
-            list.stream().forEach(addon -> addon.fileWasChanged(file));
-            active = false;
         }
     }
 }
